Improved Visual Studio warnings
 - Enable /Wall when using Visual Studio 2019
 - Fix /Wall warnings

+++ Code/Core/FileIO/FileStream.cpp
@@ -274,7 +274,7 @@ bool FileStream::IsOpen() const
     LARGE_INTEGER zeroPos, newPos;
     zeroPos.QuadPart = 0;
     VERIFY( SetFilePointerEx( (HANDLE)m_Handle, zeroPos, &newPos, FILE_CURRENT ) );
-    return newPos.QuadPart;
+    return (uint64_t)newPos.QuadPart;
 #elif defined( __APPLE__ ) || defined( __LINUX__ )
     return lseek( m_Handle, 0, SEEK_CUR );
 #else
@@ -290,7 +290,7 @@ bool FileStream::IsOpen() const
 
 #if defined( __WINDOWS__ )
     LARGE_INTEGER newPos;
-    newPos.QuadPart = pos;
+    newPos.QuadPart = (int64_t)pos;
     if ( FALSE == SetFilePointerEx( (HANDLE)m_Handle, newPos, nullptr, FILE_BEGIN ) )
     {
         return false;
