+++ Code/Tools/FBuild/FBuildCore/FLog.cpp
@@ -280,22 +280,22 @@ static FileStream * g_MonitorFileStream = nullptr;
     if ( timeTakenMinutes > 0 )
     {
         char buffer[ 8 ];
-        _itoa_s( timeTakenMinutes, buffer, 8, 10 );
+        _itoa_s( (int32_t)timeTakenMinutes, buffer, 8, 10 );
         m_ProgressText += buffer;
         m_ProgressText.Append( "m ", 2 );
     }
     char buffer[ 8 ];
-    _itoa_s( timeTakenSeconds, buffer, 8, 10 );
+    _itoa_s( (int32_t)timeTakenSeconds, buffer, 8, 10 );
     if ( timeTakenSeconds < 10 ) { m_ProgressText += '0'; }
     m_ProgressText += buffer;
     m_ProgressText += 's';
 
     // active/available jobs " (%u/%u)"
     m_ProgressText.Append( " (", 2 );
-    _itoa_s( numJobsActive, buffer, 8, 10 );
+    _itoa_s( (int32_t)numJobsActive, buffer, 8, 10 );
     m_ProgressText += buffer;
     m_ProgressText += '/';
-    _itoa_s( numJobsActive + numJobs, buffer, 8, 10 );
+    _itoa_s( (int32_t)( numJobsActive + numJobs ), buffer, 8, 10 );
     m_ProgressText += buffer;
     m_ProgressText += ')';
 
@@ -303,10 +303,10 @@ static FileStream * g_MonitorFileStream = nullptr;
     if ( FBuild::Get().GetOptions().m_AllowDistributed )
     {
         m_ProgressText.Append( "+(", 2 );
-        _itoa_s( numJobsDistActive, buffer, 8, 10 );
+        _itoa_s( (int32_t)numJobsDistActive, buffer, 8, 10 );
         m_ProgressText += buffer;
         m_ProgressText += '/';
-        _itoa_s( numJobsDistActive + numJobsDist, buffer, 8, 10 );
+        _itoa_s( (int32_t)( numJobsDistActive + numJobsDist ), buffer, 8, 10 );
         m_ProgressText += buffer;
         m_ProgressText += ')';
     }
@@ -354,7 +354,7 @@ static FileStream * g_MonitorFileStream = nullptr;
     if ( threadIndex > 0 )
     {
         char buffer[ 8 ];
-        _itoa_s( threadIndex, buffer, 8, 10 );
+        _itoa_s( (int32_t)threadIndex, buffer, 8, 10 );
         tmp += buffer;
         tmp += '>';
         if ( threadIndex < 10 )
