+++ Code/Tools/FBuild/FBuildCore/WorkerPool/Job.cpp
@@ -77,7 +77,7 @@ void Job::OwnData( void * data, size_t size, bool compressed )
         if ( m_IsLocal )
         {
             ASSERT( s_TotalLocalDataMemoryUsage >= m_DataSize );
-            AtomicSub64( &s_TotalLocalDataMemoryUsage, m_DataSize );
+            AtomicSub64( &s_TotalLocalDataMemoryUsage, (int32_t)m_DataSize );
         }
     }
 
@@ -89,7 +89,7 @@ void Job::OwnData( void * data, size_t size, bool compressed )
     // Update total memory use tracking
     if ( m_IsLocal )
     {
-        AtomicAdd64( &s_TotalLocalDataMemoryUsage, m_DataSize );
+        AtomicAdd64( &s_TotalLocalDataMemoryUsage, (int32_t)m_DataSize );
     }
 }
